fix: Move DashboardActivity to proper package
