+++ app/src/main/res/layout/app_bar_dashboard.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.softwaregroup.underthekotlintree.util.DashboardActivity">
+    tools:context="com.softwaregroup.underthekotlintree.ui.DashboardActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
