+++ app/src/main/res/layout/content_dashboard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.softwaregroup.underthekotlintree.util.DashboardActivity"
+    tools:context="com.softwaregroup.underthekotlintree.ui.DashboardActivity"
     tools:showIn="@layout/app_bar_dashboard">
 
 </android.support.constraint.ConstraintLayout>
